Store netflow stuff to come back to later
This is a holding branch until I have time to come back to netflow. The logstash
netflow codec is problematic and does not properly handle netflow data from
Cisco ASAs. To get this working, I will need to use another program to capture
the netflow from the ASA, such as nfdump.

The ES template also needs work. I need to create a raw field that stores the data
is strings so they make sense when displayed in tables. I spent a good deal of
time monkeying with the template but still haven't gotten it quite right.